docs(react): fix grammatical issue with a vs an (#15478)
