+++ packages/angular/src/generators/application/lib/update-config-files.ts
@@ -8,11 +8,11 @@ import {
   removeProjectConfiguration,
   updateJson,
 } from '@nrwl/devkit';
-import { replaceAppNameWithPath } from '@nrwl/workspace/src/utils/cli-config-utils';
 import { getRelativePathToRootTsConfig } from '@nrwl/js';
 import { E2eTestRunner, UnitTestRunner } from '../../../utils/test-runners';
 import type { NormalizedSchema } from './normalized-schema';
 import { createTsConfig } from '../../utils/create-ts-config';
+import { replaceAppNameWithPath } from '../../../utils/cli-config-utils';
 
 export function updateConfigFiles(host: Tree, options: NormalizedSchema) {
   updateTsConfigOptions(host, options);
