+++ packages/angular/src/generators/library/lib/update-project.ts
@@ -9,7 +9,7 @@ import {
   updateJson,
 } from '@nrwl/devkit';
 import { getRelativePathToRootTsConfig } from '@nrwl/js';
-import { replaceAppNameWithPath } from '@nrwl/workspace/src/utils/cli-config-utils';
+import { replaceAppNameWithPath } from '../../../utils/cli-config-utils';
 import { NormalizedSchema } from './normalized-schema';
 import { updateNgPackage } from './update-ng-package';
 import { join } from 'path';
